@@ -1,6 +1,5 @@
 {
-  "name": "trip-planner-starter-he",
-  "version": "0.1.0",
+  "name": "trip-planner",
   "private": true,
   "scripts": {
     "dev": "next dev",
@@ -9,20 +8,18 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@supabase/supabase-js": "^2.45.4",
-    "autoprefixer": "^10.4.20",
     "next": "14.2.5",
-    "postcss": "^8.4.41",
     "react": "18.3.1",
     "react-dom": "18.3.1",
-    "tailwindcss": "^3.4.10"
+    "@supabase/supabase-js": "2.45.4"
   },
   "devDependencies": {
-    "@types/node": "^18.19.129",
-    "@types/react": "^18.3.26",
-    "@types/react-dom": "^18.3.7",
-    "eslint": "^8.57.0",
-    "eslint-config-next": "14.2.5",
-    "typescript": "^5.4.5"
+    "autoprefixer": "10.4.20",
+    "postcss": "8.4.41",
+    "tailwindcss": "3.4.10",
+    "typescript": "5.6.3",
+    "@types/react": "18.3.3",
+    "eslint": "8.57.0",
+    "eslint-config-next": "14.2.5"
   }
 }
