@@ -1,2 +1,8 @@
-/** @type {import('tailwindcss').Config} */
-module.exports={content:["./app/**/*.{js,ts,jsx,tsx}","./components/**/*.{js,ts,jsx,tsx}"],theme:{extend:{colors:{brand:{50:"#fff0f5",100:"#ffd6e7",200:"#ffadd0",300:"#ff85ba",400:"#ff5da3",500:"#ff358d",600:"#cc2a70",700:"#991f54",800:"#661436",900:"#330a1b"}}}},plugins:[]};
+import type { Config } from 'tailwindcss';
+
+const config: Config = {
+  content: ['./app/**/*.{ts,tsx}', './components/**/*.{ts,tsx}', './pages/**/*.{ts,tsx}'],
+  theme: { extend: {} },
+  plugins: []
+};
+export default config;
