@@ -1,250 +1,219 @@
 'use client';
 
-import { useEffect, useMemo, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { supabase } from '@/lib/supabaseClient';
+import DrivePickerButton from '@/app/components/DrivePickerButton';
 
-type PlaceType = 'hotel' | 'transport' | 'restaurant' | 'activity';
-type Place = {
+type Trip = {
   id: string;
   name: string;
-  type: PlaceType;
-  lat: number | null;
-  lng: number | null;
-  maps_url: string | null;
-  desc_he: string | null;
+  start_date: string | null;
+  end_date: string | null;
+};
+
+type Day = {
+  id: string;
+  trip_id: string;
+  date: string;
+  order: number;
+};
+
+type Entry = {
+  id: string;
+  day_id: string;
+  type: string;
+  name: string;
+  note: string | null;
+  order: number;
   duration_min: number | null;
 };
 
-const TYPES: { key: PlaceType; label: string }[] = [
-  { key: 'hotel',      label: 'מלון' },
-  { key: 'transport',  label: 'תחבורה' },
-  { key: 'restaurant', label: 'מסעדה' },
-  { key: 'activity',   label: 'פעילות' },
-];
+export default function AdminPage() {
+  const [trips, setTrips] = useState<Trip[]>([]);
+  const [activeTrip, setActiveTrip] = useState<string | null>(null);
+  const [days, setDays] = useState<Day[]>([]);
+  const [entries, setEntries] = useState<Entry[]>([]);
+  const [newTrip, setNewTrip] = useState<Partial<Trip>>({});
+  const [loading, setLoading] = useState(false);
 
-export default function PlacesAdminPage() {
-  const [activeType, setActiveType] = useState<PlaceType>('activity');
-  const [allPlaces, setAllPlaces]     = useState<Place[]>([]);
-  const [loading, setLoading]         = useState(true);
-  const [q, setQ]                     = useState(''); // חיפוש
-  const [creating, setCreating]       = useState<Partial<Place>>({
-    name: '',
-    type: 'activity',
-    lat: null, lng: null, maps_url: '', desc_he: '', duration_min: 60,
-  });
-
-  async function load() {
-    setLoading(true);
-    const { data, error } = await supabase.from('places').select('*').order('name');
-    if (!error) setAllPlaces(data ?? []);
-    setLoading(false);
+  // טעינת טיולים
+  async function loadTrips() {
+    const { data } = await supabase.from('trips').select('*').order('start_date');
+    setTrips(data || []);
   }
 
-  useEffect(() => { void load(); }, []);
+  // טעינת ימים ופריטים
+  async function loadTripDetails(tripId: string) {
+    const { data: d } = await supabase.from('days').select('*').eq('trip_id', tripId).order('date');
+    const { data: e } = await supabase.from('day_entries').select('*').in('day_id', d?.map(x => x.id) || []).order('order');
+    setDays(d || []);
+    setEntries(e || []);
+  }
 
-  const filteredByType = useMemo(
-    () => allPlaces.filter(p => p.type === activeType),
-    [allPlaces, activeType]
-  );
+  useEffect(() => {
+    loadTrips();
+  }, []);
 
-  const filtered = useMemo(() => {
-    const term = q.trim().toLowerCase();
-    if (!term) return filteredByType;
-    return filteredByType.filter(p =>
-      (p.name ?? '').toLowerCase().includes(term) ||
-      (p.desc_he ?? '').toLowerCase().includes(term)
-    );
-  }, [filteredByType, q]);
+  useEffect(() => {
+    if (activeTrip) loadTripDetails(activeTrip);
+  }, [activeTrip]);
 
-  async function createPlace() {
-    const payload = {
-      name: (creating.name ?? '').trim(),
-      type: (creating.type as PlaceType) || 'activity',
-      lat: creating.lat ?? null,
-      lng: creating.lng ?? null,
-      maps_url: creating.maps_url || null,
-      desc_he: creating.desc_he || null,
-      duration_min: Number(creating.duration_min ?? 60),
-    };
-    if (!payload.name) return alert('שם חובה');
-    const { error } = await supabase.from('places').insert(payload);
+  async function createTrip() {
+    if (!newTrip.name) return alert('שם טיול חובה');
+    const { data, error } = await supabase.from('trips').insert({
+      name: newTrip.name,
+      start_date: newTrip.start_date,
+      end_date: newTrip.end_date
+    }).select('id').single();
     if (error) return alert(error.message);
-    setCreating({ name: '', type: activeType, lat: null, lng: null, maps_url: '', desc_he: '', duration_min: 60 });
-    await load();
+    setNewTrip({});
+    await loadTrips();
+    setActiveTrip(data.id);
   }
 
-  async function savePlace(p: Place) {
-    const { error } = await supabase
-      .from('places')
+  async function addDay() {
+    if (!activeTrip) return;
+    const date = prompt('הזן תאריך (YYYY-MM-DD)');
+    if (!date) return;
+    await supabase.from('days').insert({ trip_id: activeTrip, date, order: days.length + 1 });
+    await loadTripDetails(activeTrip);
+  }
+
+  async function addEntry(dayId: string) {
+    const type = prompt('בחר סוג (מלון, תחבורה, מסעדה, פעילות)');
+    const name = prompt('שם האטרקציה / המקום');
+    if (!name) return;
+    await supabase.from('day_entries').insert({
+      day_id: dayId,
+      type,
+      name,
+      note: '',
+      order: (entries.filter(e => e.day_id === dayId).length) + 1,
+      duration_min: 60
+    });
+    await loadTripDetails(activeTrip!);
+  }
+
+  async function saveEntry(e: Entry) {
+    const { error } = await supabase.from('day_entries')
       .update({
-        name: p.name,
-        type: p.type,
-        lat: p.lat,
-        lng: p.lng,
-        maps_url: p.maps_url,
-        desc_he: p.desc_he,
-        duration_min: p.duration_min,
+        type: e.type,
+        name: e.name,
+        note: e.note,
+        duration_min: e.duration_min
       })
-      .eq('id', p.id);
-    if (error) return alert(error.message);
-    await load();
+      .eq('id', e.id);
+    if (error) alert(error.message);
   }
 
-  async function removePlace(id: string) {
-    if (!confirm('למחוק רשומה זו?')) return;
-    await supabase.from('places').delete().eq('id', id);
-    await load();
+  async function deleteEntry(id: string) {
+    if (!confirm('למחוק פריט זה?')) return;
+    await supabase.from('day_entries').delete().eq('id', id);
+    await loadTripDetails(activeTrip!);
+  }
+
+  async function moveEntry(id: string, delta: number) {
+    const entry = entries.find(e => e.id === id);
+    if (!entry) return;
+    const sameDay = entries.filter(e => e.day_id === entry.day_id).sort((a, b) => a.order - b.order);
+    const idx = sameDay.findIndex(e => e.id === id);
+    const swap = sameDay[idx + delta];
+    if (!swap) return;
+    await supabase.from('day_entries').update({ order: swap.order }).eq('id', entry.id);
+    await supabase.from('day_entries').update({ order: entry.order }).eq('id', swap.id);
+    await loadTripDetails(activeTrip!);
   }
 
   return (
-    <main className="max-w-6xl mx-auto p-4 space-y-6">
-      <h1 className="text-2xl font-bold">ניהול אטרקציות</h1>
+    <main className="max-w-6xl mx-auto p-6 space-y-8">
+      <h1 className="text-3xl font-bold mb-4">פאנל ניהול טיולים</h1>
 
-      {/* לשוניות סוגים */}
-      <div className="flex gap-2 flex-wrap">
-        {TYPES.map(t => (
-          <button
-            key={t.key}
-            onClick={() => { setActiveType(t.key); setCreating(c => ({ ...c, type: t.key })); }}
-            className={`btn ${activeType === t.key ? '' : 'bg-gray-800'}`}
-          >
-            {t.label}
-          </button>
-        ))}
-      </div>
-
-      {/* חיפוש + הוספה */}
-      <section className="card space-y-3">
-        <div className="grid grid-cols-3 gap-2">
-          <input className="input" placeholder="חיפוש בשם/תיאור…" value={q} onChange={e => setQ(e.target.value)} />
+      {/* יצירת טיול חדש */}
+      <section className="card space-y-4">
+        <h2 className="text-xl font-semibold">צור טיול חדש</h2>
+        <div className="grid grid-cols-4 gap-3">
+          <input className="input" placeholder="שם טיול" value={newTrip.name || ''} onChange={e => setNewTrip(t => ({ ...t, name: e.target.value }))} />
+          <input className="input" type="date" value={newTrip.start_date || ''} onChange={e => setNewTrip(t => ({ ...t, start_date: e.target.value }))} />
+          <input className="input" type="date" value={newTrip.end_date || ''} onChange={e => setNewTrip(t => ({ ...t, end_date: e.target.value }))} />
+          <button className="btn" onClick={createTrip}>צור</button>
         </div>
 
-        <h2 className="font-semibold">הוסף {TYPES.find(t => t.key === activeType)?.label}</h2>
-        <div className="grid grid-cols-6 gap-2">
-          <input
-            className="input col-span-2"
-            placeholder="שם"
-            value={(creating.name as any) || ''}
-            onChange={e => setCreating(r => ({ ...r, name: e.target.value }))}
-          />
-          <input
-            className="input"
-            placeholder="lat"
-            value={(creating.lat as any) ?? ''}
-            onChange={e => setCreating(r => ({ ...r, lat: e.target.value === '' ? null : Number(e.target.value) }))}
-          />
-          <input
-            className="input"
-            placeholder="lng"
-            value={(creating.lng as any) ?? ''}
-            onChange={e => setCreating(r => ({ ...r, lng: e.target.value === '' ? null : Number(e.target.value) }))}
-          />
-          <input
-            className="input col-span-2"
-            placeholder="קישור מפה (לא חובה)"
-            value={(creating.maps_url as any) || ''}
-            onChange={e => setCreating(r => ({ ...r, maps_url: e.target.value }))}
-          />
-          <textarea
-            className="input col-span-5"
-            placeholder="תיאור"
-            value={(creating.desc_he as any) || ''}
-            onChange={e => setCreating(r => ({ ...r, desc_he: e.target.value }))}
-          />
-          <input
-            className="input"
-            type="number"
-            placeholder="משך (דקות)"
-            value={(creating.duration_min as any) ?? 60}
-            onChange={e => setCreating(r => ({ ...r, duration_min: Number(e.target.value || 60) }))}
-          />
-          <button className="btn col-span-2" onClick={createPlace}>שמור</button>
+        <div className="flex flex-wrap gap-2">
+          {trips.map(t => (
+            <button
+              key={t.id}
+              className={`btn ${activeTrip === t.id ? '' : 'bg-gray-800'}`}
+              onClick={() => setActiveTrip(t.id)}
+            >
+              {t.name}
+            </button>
+          ))}
         </div>
       </section>
 
-      {/* טבלה */}
-      <section className="card">
-        {loading ? (
-          <div className="p-4 text-gray-500">טוען…</div>
-        ) : (
-          <div className="overflow-auto">
-            <table className="w-full text-sm">
-              <thead>
-                <tr className="text-right text-gray-600">
-                  <th className="p-2">#</th>
-                  <th className="p-2">שם</th>
-                  <th className="p-2">lat</th>
-                  <th className="p-2">lng</th>
-                  <th className="p-2">מפה</th>
-                  <th className="p-2">משך</th>
-                  <th className="p-2">תיאור</th>
-                  <th className="p-2">פעולות</th>
-                </tr>
-              </thead>
-              <tbody>
-                {filtered.map((p, i) => (
-                  <tr key={p.id} className="border-t">
-                    <td className="p-2">{i + 1}</td>
-                    <td className="p-2">
-                      <input className="input w-56" value={p.name} onChange={e => (p.name = e.target.value)} />
-                    </td>
-                    <td className="p-2">
-                      <input
-                        className="input w-28"
-                        value={p.lat ?? ''}
-                        onChange={e => (p.lat = e.target.value === '' ? null : Number(e.target.value))}
-                      />
-                    </td>
-                    <td className="p-2">
-                      <input
-                        className="input w-28"
-                        value={p.lng ?? ''}
-                        onChange={e => (p.lng = e.target.value === '' ? null : Number(e.target.value))}
-                      />
-                    </td>
-                    <td className="p-2">
-                      <input className="input w-56" value={p.maps_url ?? ''} onChange={e => (p.maps_url = e.target.value)} />
-                    </td>
-                    <td className="p-2">
-                      <input
-                        className="input w-20"
-                        type="number"
-                        value={p.duration_min ?? 60}
-                        onChange={e => (p.duration_min = Number(e.target.value || 60))}
-                      />
-                    </td>
-                    <td className="p-2">
-                      <textarea className="input w-72" value={p.desc_he ?? ''} onChange={e => (p.desc_he = e.target.value)} />
-                    </td>
-                    <td className="p-2">
-                      <div className="flex gap-2">
-                        <button className="btn" onClick={() => savePlace(p)}>שמור</button>
-                        <button className="btn bg-gray-800" onClick={() => removePlace(p.id)}>מחק</button>
-                      </div>
-                    </td>
+      {/* ימים */}
+      {activeTrip && (
+        <>
+          <button className="btn" onClick={addDay}>הוסף יום</button>
+
+          {days.map(day => (
+            <section key={day.id} className="card space-y-3">
+              <h3 className="text-lg font-semibold">
+                יום {new Date(day.date).toLocaleDateString('he-IL')}
+              </h3>
+
+              <button className="btn" onClick={() => addEntry(day.id)}>הוסף פריט</button>
+
+              <table className="w-full text-right text-sm">
+                <thead>
+                  <tr className="text-gray-600">
+                    <th>#</th>
+                    <th>סוג</th>
+                    <th>שם</th>
+                    <th>משך (דק')</th>
+                    <th>הערה</th>
+                    <th>קבצים</th>
+                    <th>פעולות</th>
                   </tr>
-                ))}
-                {filtered.length === 0 && (
-                  <tr><td className="p-2 text-gray-500" colSpan={8}>אין רשומות</td></tr>
-                )}
-              </tbody>
-            </table>
-          </div>
-        )}
-      </section>
+                </thead>
+                <tbody>
+                  {entries.filter(e => e.day_id === day.id).sort((a, b) => a.order - b.order).map((e, i) => (
+                    <tr key={e.id} className="border-t">
+                      <td>{i + 1}</td>
+                      <td>
+                        <input className="input w-24" value={e.type || ''} onChange={ev => (e.type = ev.target.value)} />
+                      </td>
+                      <td>
+                        <input className="input w-52" value={e.name || ''} onChange={ev => (e.name = ev.target.value)} />
+                      </td>
+                      <td>
+                        <input className="input w-20" type="number" value={e.duration_min || 60} onChange={ev => (e.duration_min = Number(ev.target.value))} />
+                      </td>
+                      <td>
+                        <textarea className="input w-60" value={e.note || ''} onChange={ev => (e.note = ev.target.value)} />
+                      </td>
+                      <td>
+                        <DrivePickerButton entryId={e.id} />
+                      </td>
+                      <td>
+                        <div className="flex gap-2">
+                          <button className="btn" onClick={() => saveEntry(e)}>שמור</button>
+                          <button className="btn bg-gray-700" onClick={() => deleteEntry(e.id)}>מחק</button>
+                          <button className="btn bg-gray-500" onClick={() => moveEntry(e.id, -1)}>▲</button>
+                          <button className="btn bg-gray-500" onClick={() => moveEntry(e.id, 1)}>▼</button>
+                        </div>
+                      </td>
+                    </tr>
+                  ))}
+                  {entries.filter(e => e.day_id === day.id).length === 0 && (
+                    <tr><td colSpan={7} className="text-gray-400">אין פריטים ביום זה</td></tr>
+                  )}
+                </tbody>
+              </table>
+            </section>
+          ))}
+        </>
+      )}
     </main>
   );
 }
-
-/* מחלקות עזר אם אינן קיימות ב-globals.css
-.input { @apply w-full rounded-xl border border-gray-300/60 bg-white/60 px-3 py-2 text-right; }
-.btn   { @apply rounded-xl bg-pink-600 px-3 py-2 text-white hover:opacity-90; }
-.card  { @apply rounded-2xl border border-gray-200/60 bg-white/70 p-4 shadow-sm; }
-*/
-import Link from 'next/link';
-
-// איפשהו ליד הכותרת:
-<div className="flex gap-2">
-  <h1 className="text-2xl font-bold">פאנל ניהול</h1>
-  <Link href="/admin/places" className="btn">ניהול אטרקציות</Link>
-</div>
