@@ -1,1 +1,3 @@
-export async function GET(){return new Response('ok');}
+export async function GET() {
+  return new Response('ok');
+}
