@@ -1,6 +1,15 @@
-export const metadata={title:'מתכנן טיולים',description:'MVP חינמי בעברית עם פאנל עריכה'};
-import './../styles/globals.css';
-import React from 'react';
-export default function RootLayout({children}:{children:React.ReactNode}){
- return(<html lang='he' dir='rtl'><body>{children}</body></html>);
+import './globals.css';
+import '../styles/globals.css';
+
+export const metadata = {
+  title: 'מתכנן טיולים',
+  description: 'Trip planner admin'
+};
+
+export default function RootLayout({ children }: { children: React.ReactNode }) {
+  return (
+    <html lang="he" dir="rtl">
+      <body className="bg-rose-50 text-neutral-900">{children}</body>
+    </html>
+  );
 }
