@@ -1,3 +1,6 @@
 /** @type {import('next').NextConfig} */
-const nextConfig={reactStrictMode:true,experimental:{appDir:true}};
-export default nextConfig;
+const nextConfig = {
+  reactStrictMode: true
+};
+
+module.exports = nextConfig;
