@@ -1,1 +1,6 @@
-module.exports={plugins:{tailwindcss:{},autoprefixer:{},},};
+module.exports = {
+  plugins: {
+    tailwindcss: {},
+    autoprefixer: {}
+  }
+};
